+++ source/robot_model/src/Model.cpp
@@ -393,7 +393,8 @@ void Model::check_inverse_velocity_arguments(const std::vector<state_representat
 state_representation::JointVelocities
 Model::inverse_velocity(const std::vector<state_representation::CartesianTwist>& cartesian_twists,
                         const state_representation::JointPositions& joint_positions,
-                        const std::vector<std::string>& frames) {
+                        const std::vector<std::string>& frames,
+                        const double dls_lambda) {
   // sanity check
   this->check_inverse_velocity_arguments(cartesian_twists, joint_positions, frames);
 
@@ -411,19 +412,37 @@ Model::inverse_velocity(const std::vector<state_representation::CartesianTwist>&
   dX.tail(6) = cartesian_twists.back().data();
   jacobian.bottomRows(6) = this->compute_jacobian(joint_positions, frames.back()).data();
 
-  // solve a linear system
-  return state_representation::JointVelocities(joint_positions.get_name(),
+  if (dls_lambda == 0.0){
+    return state_representation::JointVelocities(joint_positions.get_name(),
                                                joint_positions.get_names(),
                                                jacobian.colPivHouseholderQr().solve(dX));
+  }
+
+  // add damped least square term
+  if (jacobian.rows() > jacobian.cols()){
+    Eigen::MatrixXd j_prime = jacobian.transpose() * jacobian + 
+                dls_lambda * dls_lambda * Eigen::MatrixXd::Identity(jacobian.cols(), jacobian.cols());
+    return state_representation::JointVelocities(joint_positions.get_name(),
+                                               joint_positions.get_names(),
+                                               j_prime.colPivHouseholderQr().solve(jacobian.transpose() * dX));
+  } else {
+    Eigen::MatrixXd j_prime = jacobian * jacobian.transpose() + 
+                dls_lambda * dls_lambda * Eigen::MatrixXd::Identity(jacobian.rows(), jacobian.rows());
+    return state_representation::JointVelocities(joint_positions.get_name(),
+                                               joint_positions.get_names(),
+                                               jacobian.transpose() * j_prime.colPivHouseholderQr().solve(dX));
+  }
 }
 
 state_representation::JointVelocities Model::inverse_velocity(const state_representation::CartesianTwist& cartesian_twist,
                                                               const state_representation::JointPositions& joint_positions,
-                                                              const std::string& frame) {
+                                                              const std::string& frame,
+                                                              const double dls_lambda) {
   std::string actual_frame = frame.empty() ? this->robot_model_.frames.back().name : frame;
   return this->inverse_velocity(std::vector<state_representation::CartesianTwist>({cartesian_twist}),
                                 joint_positions,
-                                std::vector<std::string>({actual_frame}));
+                                std::vector<std::string>({actual_frame}),
+                                dls_lambda);
 }
 
 state_representation::JointVelocities